pyproject.toml
[tool.poetry]
name = "nornir3_demo"
version = "1.0.0"
description = "This is a demo project for nornir3"
authors = ["David Barroso <dbarrosop@dravetech.com>"]
license = "Apache-2.0"

[tool.poetry.plugins."nornir.plugins.inventory"]
"ACMEInventory" = "nornir3_demo.plugins.inventory.acme:ACMEInventory"

[tool.poetry.plugins."nornir.plugins.connections"]
"acmeos" = "nornir3_demo.plugins.connections.acmeos:AcmeOS"

[tool.poetry.plugins."nornir.plugins.runners"]
"DCAwareRunner" = "nornir3_demo.plugins.runners.dc_aware:DCAwareRunner"

[tool.poetry.dependencies]
python = "^3.6"
nornir = { version = "~3.0.0a4", allow-prereleases = true }
rich = "^1.3.1"
prometheus_client = "^0.8.0"
flask = "^1.1.2"
uwsgi = "^2.0.18"
"ruamel.yaml" = "^0.16.10"

[tool.poetry.dev-dependencies]
black = { version = "19.10b0", allow-prereleases = true }
mypy = "*"
pylama = "^7.7.1"
pytest = "*"

[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.masonry.api"
